requirements.txt
aiogram==3.13.1
aiohttp==3.10.11
python-dotenv==1.0.1
requests==2.32.3